Nests now only go in unoccupied spaces!
The script now successfully checks the desired area before spawning a nest, and will create a new location if it is currently occupied.

However, as there is little space to work with, the spawn range has been greatly increased in the x value to prevent a stack overflow.

The next task to be done is to have nest branches move in the same direction so they appear more naturally into less filled areas.