Started work on player map navigation!
Added the Nest Navigation script.

Also added the functionality to the map script to have each nest store the following nest as a gameobject. This hasn't yet been added to nests in a branch, which is the next task.

The navigation script will function by saving the current nest it's on and entering it's battle, and once it's complete, moving to the next nest that's saved as a gameobject in the current nest.