Map screen now takes player to fight
Added the ability to move to a fight once the target distance is met to NestNavigator. Also added a DataStorer, which reads the desired monster from the nest and carries it into the fight scene using DontDestroyOnLoad.

Also added a DataReader, which spawns a monster based on the monster saved in DataStorer.

Accidentally messed with files, health number no longer goes down in fight scene.

Need to add ability for fight scene to end and return to nest once a winner is determined.