Game now returns to map on enemy defeat
The map is now saved by Dataholder. NestNavigator sets its 'map' gameobject as the current map, and de-activated it before loading in a new scene.

It's then re-activated by the NestReactivator in the new NestMapless scene.

This requires the whole map to be loaded in during the fight scene using DontDestroyOnLoad, which could slow the game down. Try to find alternative during optimisation.

Next task is to move onto next nest once one is complete. Then to create a main menu and return to it if the player fails.